test(api): isolate embedding worker Compose contract
